[sdk] split out verifying client logic into separate files
