+++ sdk/client/src/verifying_client/blocking.rs
@@ -1,0 +1,7 @@
+// Copyright (c) The Diem Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+// TODO(philiphayes): fill out
+
+#[allow(dead_code)]
+pub struct BlockingVerifyingClient;
