+++ sdk/client/src/verifying_client/mod.rs
@@ -2,4 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod client;
-pub use client::{InMemoryStorage, Storage, VerifyingClient};
+pub use client::VerifyingClient;
+
+pub mod blocking;
+pub use blocking::BlockingVerifyingClient;
+
+pub mod state_store;
+pub use state_store::{InMemoryStorage, Storage};
+
+mod methods;
